🚚 : 스토리북 템플릿 위치 수정
